pip_requirements.txt

#
# pybtex.
#
pybtex >= 0.19

#
# arxiv2bib.
#
arxiv2bib >= 1.0.2

#
# My own pylatexenc.
#
pylatexenc >= 2.0

#
# `future` for both python2 & python3 support
#
future

#
# `appdirs` to find where to store config files and caches
#
appdirs

#
# `PyGithub` and `requests`
#
PyGithub
requests[security]

#
# `bs4` and `requests` for citeinspirehep filter
#
bs4
#requests
